requirements.txt
streamlit
langchain
langchain-groq
langchain-community
sentence-transformers
faiss-cpu
youtube-transcript-api
python-dotenv
